src/pyproject.toml
[project]
name = "x-mcp"
version = "0.1.0"
description = "MCP server providing a tool to fetch recent tweets for a given X (Twitter) username"
authors = [ { name = "Your Name", email = "you@example.com" } ]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["mcp", "model context protocol", "x", "twitter", "agent", "tool"]
dependencies = [
    "mcp[cli]>=1.14.0",
    "requests>=2.32.5",
    "python-dotenv>=1.0.1",
    "langchain-openai>=0.3.33",
    "mcp-use>=1.3.10",
    "azure-identity>=1.25.0",
    "argparse>=1.4.0",
    "fastapi>=0.112.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.9",
]
